Need clarification on "a brief summary"
Specifically, I need to know why the metrics from the "a brief summary" section of the Jupyter Notebook do not match the actual results of the graphs. Additionally, why are we comparing experiment 1 to experiment 3 if experiment 3 couldn't run successfully, according to the comments in the code chunks and the conclusion. I mentioned more of my issues/confusion in the group chat. 6 AM now so I need to sleep.